requirements.txt
aiogram==3.4.1
flask
google-generativeai
supabase
python-dotenv
requests
pytest
psycopg2
dotenv